requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.12

# Data Science and ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Computer Vision
pillow>=10.0.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress and Logging
tqdm>=4.66.0

# AI Integration (Optional - for advanced reporting)
google-generativeai>=0.3.0

# PDF Generation (Optional - for reports)
reportlab>=4.0.0

# Cloud Storage (Optional - for Firebase integration)
firebase-admin>=6.2.0